Move parHIP implementation to core/parhip.c
