+++ core/makefile.template
@@ -24,6 +24,7 @@ USR_LFLAGS=
 MPI=0 
 VISIT=0
 DPROCMAP=0
+
 PARMETIS=0
 ZOLTAN=0
 ZOLTAN2=0
@@ -105,6 +106,10 @@ endif
 
 ifeq ($(ZOLTAN2), 1)
 	CORE := $(CORE) zoltan2.c.o zoltan2.cpp.o
+endif
+
+ifeq ($(PARHIP), 1)
+	CORE := $(CORE) parhip.o
 endif
 
 CORE := $(CORE) partitioner-weak.o
@@ -262,6 +267,7 @@ $(OBJDIR)/partitioner-weak.o :$S/core/partitioner-weak.c;  $(CC) -c $(cFL2) $(GS
 $(OBJDIR)/parmetis.o         :$S/core/parmetis.c;          $(CC) -c $(cFL2) $(GSLIB_IFLAGS) $(PART_C_IFLAGS) $< -o $@
 $(OBJDIR)/zoltan.o           :$S/core/zoltan.c;            $(CC) -c $(cFL2) $(GSLIB_IFLAGS) $(PART_C_IFLAGS) $< -o $@
 $(OBJDIR)/zoltan2.c.o        :$S/core/zoltan2.c;           $(CC) -c $(cFL2) $(GSLIB_IFLAGS) $< -o $@
+$(OBJDIR)/parhip.o           :$S/core/parhip.c;            $(CC) -c $(cFL2) $(GSLIB_IFLAGS) $(PART_C_IFLAGS) $< -o $@
 
 # CXX Files ##################################################################################
 $(OBJDIR)/zoltan2.cpp.o :$S/core/zoltan2.cpp; $(CXX) -c $(cxxFL2) $(GSLIB_IFLAGS) $(PARMETIS_IFLAGS) $(ZOLTAN2_IFLAGS) $< -o $@
